+++ RazorSvelte/App/lib/theme.ts
@@ -1,28 +1,12 @@
 import { writable } from "svelte/store";
 import { themeKey } from "$lib/config";
 
-const dark = "dark";
-const light = "light";
-
-const html = document.getElementsByTagName("html")[0] as HTMLHtmlElement;
-const defaultTheme: string = html?.dataset.bsTheme as string;
-
-export const isDarkTheme = writable<boolean>(defaultTheme == dark);
+export const dark = "dark";
+export const light = "light";
 export const themes: ThemesType[] = [light, dark];
-export const theme = writable<ThemesType>(defaultTheme == dark ? dark : light);
+export const theme = writable<ThemesType>(localStorage.getItem(themeKey) == dark ? dark : light);
 
-theme.subscribe((theme) => isDarkTheme.set(theme == dark));
-
-if (html) {
-    isDarkTheme.subscribe((isDark) => {
-        const d = new Date();
-        d.setFullYear(d.getFullYear() + 10);
-        if (!isDark) {
-            html.dataset.bsTheme = light;
-            document.cookie = `${themeKey}=${light}; expires=${d.toUTCString()}; path=/`;
-        } else {
-            html.dataset.bsTheme = dark;
-            document.cookie = `${themeKey}=${dark}; expires=${d.toUTCString()}; path=/`;
-        }
-    });
-}
+theme.subscribe((value) => {
+    document.getElementsByTagName("html")[0].dataset.bsTheme = value;
+    localStorage.setItem(themeKey, value);
+});